Set focus on the password field
Previously when the dialog opened the focus was on the OK button and
one had to click on the KLineEdit widget to enter the password. With
this change the foucs is on the KLineEdit widget instead and one can
start to enter the password right away.

In addition to this make sure the dialog doesn't steal focus. Other-
wise it might interrupt text input in some other window when it opens.

REVIEW: 108991
BUG: 312338