Add runtime-only dependency to ktp-auth-handler
There is already a big warning in ktp-auth-handler about the need of
qca-ossl plugin during runtime. But according to packagers this can
easily get missed and is not machine parseable and I was suggested to
rather do it this way.

Technically this does not add any new dependency, it just marks existing
runtime dependencies in a better way and makes life easier for
packagers.

REVIEW: 122868